removing min staff number
