+++ client/src/main/scala/drt/client/services/SPACircuit.scala
@@ -179,9 +179,7 @@ case class RootModel(applicationVersion: Pot[ClientServerVersions] = Empty,
                      slaConfigs: Pot[SlaConfigs] = Empty,
                      showFeedbackBanner: Pot[Boolean] = Empty,
                      userSelectedPlanningTimePeriod: Pot[Int] = Empty,
-                     flightHighlight: FlightHighlight = FlightHighlight(false, false, false, Seq.empty, Set.empty[Country], ""),
-                     minStaff: Pot[TerminalMinStaff] = Empty,
-                     showMinStaffSuccess: Pot[Boolean] = Empty
+                     flightHighlight: FlightHighlight = FlightHighlight(false, false, false, Seq.empty, Set.empty[Country], "")
                     )
 
 object PollDelay {
@@ -262,7 +260,6 @@ trait DrtCircuit extends Circuit[RootModel] with ReactConnector[RootModel] {
       new UserFeedbackBannerHandler(zoomRW(_.showFeedbackBanner)((m, v) => m.copy(showFeedbackBanner = v))),
       new UserPreferencesHandler(zoomRW(_.userSelectedPlanningTimePeriod)((m, v) => m.copy(userSelectedPlanningTimePeriod = v))),
       new FlightHighlightHandler(zoomRW(_.flightHighlight)((m, v) => m.copy(flightHighlight = v))),
-      new MinStaffHandler(zoomRW(_.minStaff)((m, v) => m.copy(minStaff = v))),
     )
     composedHandlers
   }
