+++ server/src/main/scala/services/staffing/StaffMinutesChecker.scala
@@ -14,7 +14,6 @@ case class StaffMinutesChecker(now: () => SDateLike,
                                staffingUpdateRequestQueue: ActorRef,
                                forecastMaxDays: Int,
                                airportConfig: AirportConfig,
-                               setConfiguredMinimumStaff: (Terminal, LocalDate) => Future[Done],
                               )
                               (implicit ec: ExecutionContext) {
   private val log = LoggerFactory.getLogger(getClass)
@@ -23,12 +22,9 @@ case class StaffMinutesChecker(now: () => SDateLike,
     (forecastMaxDays - 2 until forecastMaxDays).foreach { daysInFuture =>
       val date = now().addDays(daysInFuture).toLocalDate
       airportConfig.terminals.foreach { terminal: Terminal =>
-        setConfiguredMinimumStaff(terminal, date)
-          .foreach { _ =>
             log.info(s"Requesting staff minutes calculation for $terminal on $date")
             val request = TerminalUpdateRequest(terminal, date)
             staffingUpdateRequestQueue ! request
-          }
       }
     }
   }
