+++ server/src/test/scala/controllers/application/StaffingControllerSpec.scala
@@ -15,7 +15,7 @@ import uk.gov.homeoffice.drt.db.tables.{PortTerminalConfig, PortTerminalShiftCon
 import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Terminals.{T1, Terminal}
 import uk.gov.homeoffice.drt.ports.config.Lhr
-import uk.gov.homeoffice.drt.service.staffing.{FixedPointsService, MinimumStaff, MinimumStaffingService, ShiftsService, StaffMovementsService, StaffShiftFormService}
+import uk.gov.homeoffice.drt.service.staffing.{FixedPointsService, ShiftsService, StaffMovementsService, StaffShiftFormService}
 import uk.gov.homeoffice.drt.time.{LocalDate, SDate, SDateLike}
 import upickle.default.write
 
@@ -194,45 +194,6 @@ class StaffingControllerSpec extends PlaySpec with BeforeAndAfterEach {
     }
   }
 
-  "getMinimumStaffing" should {
-    "return the minimum staffing from the mock service as json" in {
-      val authHeader = Headers("X-Forwarded-Groups" -> "LHR")
-      val result = controller
-        .getMinimumStaff("t2")
-        .apply(FakeRequest(method = "GET", uri = "", headers = authHeader, body = AnyContentAsEmpty))
-
-      status(result) must ===(OK)
-      contentAsString(result) must ===(write(MinimumStaff(10)))
-    }
-  }
-
-  "saveMinimumStaffing" should {
-    "return Accepted given the correct permission and a valid value" in {
-      val authHeader = Headers("X-Forwarded-Groups" -> "staff:edit,LHR")
-      val result = controller
-        .saveMinimumStaff("t2")
-        .apply(FakeRequest(method = "POST", uri = "", headers = authHeader, body = AnyContentAsText(write(MinimumStaff(10)))))
-
-      status(result) must ===(ACCEPTED)
-    }
-    "return BadRequest given the correct permission but an invalid value" in {
-      val authHeader = Headers("X-Forwarded-Groups" -> "staff:edit,LHR")
-      val result = controller
-        .saveMinimumStaff("t2")
-        .apply(FakeRequest(method = "POST", uri = "", headers = authHeader, body = AnyContentAsEmpty))
-
-      status(result) must ===(BAD_REQUEST)
-    }
-    "return Forbidden given incorrect permissions" in {
-      val authHeader = Headers("X-Forwarded-Groups" -> "LHR")
-      val result = controller
-        .saveMinimumStaff("t2")
-        .apply(FakeRequest(method = "POST", uri = "", headers = authHeader, body = AnyContentAsEmpty))
-
-      status(result) must ===(FORBIDDEN)
-    }
-  }
-
   "when called without necessary role header, the endpoints" should {
     "return Forbidden for getShifts" in {
       val result = controller
@@ -306,20 +267,12 @@ class StaffingControllerSpec extends PlaySpec with BeforeAndAfterEach {
       MockShiftsService(shifts),
       MockFixedPointsService(fixedPoints),
       MockStaffMovementsService(movements),
-      MockMinimumStaffingService(),
       MockStaffShiftFormService()
     )
 
   private def newDrtInterface(): DrtSystemInterface = {
     new TestDrtModule(Lhr.config).provideDrtSystemInterface
   }
-}
-
-case class MockMinimumStaffingService() extends MinimumStaffingService {
-  override val getTerminalConfig: Terminal => Future[Option[PortTerminalConfig]] = _ => Future.successful(Option(PortTerminalConfig(PortCode("STN"), T1, Option(10), 0)))
-
-  override def setMinimum(terminal: Terminal, newMinimum: Option[Int]): Future[ShiftAssignments] =
-    Future.successful(ShiftAssignments(Seq()))
 }
 
   case class MockStaffShiftFormService() extends StaffShiftFormService{
